MoulinRouge/bots/eventbrite/requirements.txt
asn1crypto==0.24.0
attrs==19.3.0
Automat==0.8.0
certifi==2019.11.28
cryptography==2.8
cssselect==1.1.0
ipaddress==1.0.23
kafka-python
lxml==4.4.2
pycrypto==2.6.1
PyDispatcher==2.0.5
pyOpenSSL==19.1.0
queuelib==1.5.0
requests==2.22.0
Scrapy==1.8.0
urllib3==1.25.7
w3lib==1.21.0
zope.interface==4.7.1
geopy==2.0.0
gender-guesser==0.4.0
pymongo==3.10.1
python-dateutil==2.7.3
timestring==1.6.4